src/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.1
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.0
frozenlist==1.3.1
greenlet==1.1.3
gunicorn==20.1.0
idna==2.8
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
more-itertools==8.14.0
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pathspec==0.10.1
platformdirs==2.5.2
pluggy==0.13.1
psycopg2-binary==2.9.4
py==1.11.0
pyee==9.0.4
pyparsing==3.0.9
pytest==4.6.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.22.0
six==1.16.0
slack-sdk==3.19.1
slackclient==2.9.4
slackeventsapi==3.0.1
sleeper-api-wrapper==1.0.7
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0
